add board method get, post, put, delete And add swagger ui

+++ src/main/java/org/shop/api/moduls/board/Board.java
@@ -13,7 +13,7 @@ import java.time.LocalDateTime;
 @RequiredArgsConstructor
 public class Board implements Serializable {
 
-    private Long boardNo;
+    private Integer boardNo;
     @NonNull
     private String title;
     private String writer;
